requirements.txt
pandas
requests
matplotlib
seaborn
plotly
openpyxl
tqdm
python-dotenv